Add files via upload
An HC-SR04 ultrasonic distance sensor actually consists of 2 transducers, one of which acts as a transmitter that 
converts the electrical signal into 40 KHz ultrasonic sound pulses. The other acts as a receiver and listens for the 
transmitted pulses which echo back. 
When the receiver gets back the sounds, it produces an output pulse whose width is proportional to the distance of the 
object in front. This sensor provides excellent non-contact range detection between 2 cm to 400 cm (~13 feet) with an 
accuracy of 3 mm. 
As it operates on 5 volts, it could be directly connected to an Arduino or any other 5V logic microcontroller.